refactor(context): extract selection pack helpers
